Rename the packages.
